--- medical_shop/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MedicalShopConfig(AppConfig):
-    name = 'medical_shop'
