+++ stomatology/medical_appointments/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Appointment
+
+admin.site.register(Appointment)
