+++ stomatology/medical_appointments/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class MedicalAppointmentsConfig(AppConfig):
+    name = 'medical_appointments'
