+++ stomatology/medical_appointments/templates/medical_appointments/appointment_list.html
