+++ stomatology/stomatology/urls.py
@@ -4,13 +4,7 @@ from users import views as user_views
 from django.contrib.auth import views as auth_views
 from django.conf import settings
 from django.conf.urls.static import static
-from users.views import (
-    CreateAppointment,
-    ListMyAppointments,
-    ListDoctorAppointments,
-    AjaxHandlerView,
-)
-from users.decorators import doctor_only, unauthenticated_user, chip_number_needed
+from users.decorators import unauthenticated_user, chip_number_needed
 from django.contrib.auth.decorators import login_required
 
 urlpatterns = [
@@ -32,24 +26,8 @@ urlpatterns = [
         name="logout",
     ),
     path("profile/", user_views.profile, name="profile"),
-    path("update_profile/", user_views.update_profile, name="update_profile"),
-    path(
-        "appointment/",
-        chip_number_needed(CreateAppointment.as_view()),
-        name="appointment",
-    ),
-    path("my_appointments/", ListMyAppointments.as_view(), name="my_appointments"),
-    path(
-        "doctor_appointments/",
-        doctor_only(ListDoctorAppointments.as_view()),
-        name="doctor_appointments",
-    ),
-    path(
-        "doctor_appointments/<int:appointment_id>/",
-        user_views.doctor_update_appointment,
-        name="doctor_appointment_update",
-    ),
-    path("ajax/appointment", AjaxHandlerView.as_view(), name="ajax_appointment"),
+    path("profile/update", user_views.update_profile, name="update_profile"),
+    path("appointments/", include("medical_appointments.urls")),
 ]
 
 if settings.DEBUG:
