+++ stomatology/users/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Appointment, Doctor, Profile
+from .models import Doctor, Profile
 
 admin.site.register(Profile)
 admin.site.register(Doctor)
-admin.site.register(Appointment)
